# PERF double-groupby-5: fast checking the control bits on iterator for the cases in which the delta is 0