Fast checking the control bits for the cases in which the delta is 0
This avoids the call to expensive `readInteger` and `readFloat` functions

master results on oss-standalone:
```
2021-12-01 10:30:16,184 INFO Benchmark duration 35 secs.
2021-12-01 10:30:16,184 WARNING Benchmark duration of 35 secs is bellow the considered minimum duration for a stable run (60 secs).
2021-12-01 10:30:16,184 INFO Extracting the benchmark results
2021-12-01 10:30:16,184 INFO stdout: None
2021-12-01 10:30:16,184 INFO stderr: None
# Results for tsbs-scale100_double-groupby-5 test-case on oss-standalone topology
|            Metric JSON Path            |Metric Value|
|----------------------------------------|-----------:|
|Totals.overallQuantiles.all_queries.q0  |       0.000|
|Totals.overallQuantiles.all_queries.q50 |    2296.447|
|Totals.overallQuantiles.all_queries.q95 |    2916.735|
|Totals.overallQuantiles.all_queries.q99 |    3180.031|
|Totals.overallQuantiles.all_queries.q100|    4415.231|
|Totals.overallQueryRates.all_queries    |      28.306|
```

branch results:
```
2021-12-01 14:14:49,897 INFO Benchmark duration 31 secs.
2021-12-01 14:14:49,898 WARNING Benchmark duration of 31 secs is bellow the considered minimum duration for a stable run (60 secs).
2021-12-01 14:14:49,898 INFO Extracting the benchmark results
2021-12-01 14:14:49,898 INFO stdout: None
2021-12-01 14:14:49,898 INFO stderr: None
# Results for tsbs-scale100_double-groupby-5 test-case on oss-standalone topology
|            Metric JSON Path            |Metric Value|
|----------------------------------------|-----------:|
|Totals.overallQuantiles.all_queries.q0  |       0.000|
|Totals.overallQuantiles.all_queries.q50 |    2045.887|
|Totals.overallQuantiles.all_queries.q95 |    2646.143|
|Totals.overallQuantiles.all_queries.q99 |    2948.351|
|Totals.overallQuantiles.all_queries.q100|    4040.447|
|Totals.overallQueryRates.all_queries    |      31.644|
2021-12-01 14:14:49,942 INFO Tearing down setup oss-standalone
2021-12-01 14:14:49,945 INFO Tear-down completed
```