@filipecosta90 2021-12-01
On overall, this represents 5% latency reduction, which can reach up to 11% on some of the cases.

Used command to produce the comparison tables ( adjust setup type and name )
```
TRIGGERING_ENV=circleci redisbench-admin compare     --baseline-branch master --comparison-branch perf.compressed.iterator --github_repo RedisTimeSeries --github_org RedisTimeSeries  --metric_name "Totals.overallQuantiles.all_queries.q50" --deployment_name oss-cluster-30-primaries --deployment_type oss-cluster --metric_mode "lower-better"
```

# standalone

## Comparison between master and perf.compressed.iterator for metric: Totals.overallQuantiles.all_queries.q50
|             Test Case             |Baseline value|Comparison Value|% change (lower-better)|
|-----------------------------------|-------------:|---------------:|----------------------:|
|tsbs-scale100_cpu-max-all-1        |        45.549|          43.259|                   5.29|
|tsbs-scale100_cpu-max-all-8        |       275.439|         255.847|                   7.66|
|tsbs-scale100_double-groupby-1     |       591.039|         564.351|                   4.73|
|tsbs-scale100_double-groupby-5     |      2879.871|        2873.471|                   0.22|
|tsbs-scale100_double-groupby-all   |      5623.295|        5634.303|                  -0.20|
|tsbs-scale100_groupby-orderby-limit|        63.457|          63.411|                   0.07|
|tsbs-scale100_high-cpu-1           |       154.743|         151.551|                   2.11|
|tsbs-scale100_high-cpu-all         |     16672.255|       16347.135|                   1.99|
|tsbs-scale100_lastpoint            |       113.135|         113.715|                  -0.51|
|tsbs-scale100_single-groupby-1-1-1 |         2.492|           2.681|                  -7.05 (ignore it due to unstable query...) |
|tsbs-scale100_single-groupby-1-1-12|        15.311|          15.310|                   0.01|
|tsbs-scale100_single-groupby-1-8-1 |        26.068|          26.219|                  -0.58|
|tsbs-scale100_single-groupby-5-1-1 |        44.231|          44.269|                  -0.09|
|tsbs-scale100_single-groupby-5-1-12|       146.479|         145.767|                   0.49|
|tsbs-scale100_single-groupby-5-8-1 |        91.791|          91.323|                   0.51|


# 05-primaries

## Comparison between master and perf.compressed.iterator for metric: Totals.overallQuantiles.all_queries.q50
|                 Test Case                 |Baseline value|Comparison Value|% change (lower-better)|
|-------------------------------------------|-------------:|---------------:|----------------------:|
|tsbs-scale100-cpu-max-all-1@4139rps        |        11.621|          11.309|                   2.76|
|tsbs-scale100-cpu-max-all-8@588rps         |        75.107|          71.715|                   4.73|
|tsbs-scale100-double-groupby-1@324rps      |       113.955|         109.779|                   3.80|
|tsbs-scale100-double-groupby-5@70rps       |       572.703|         549.503|                   4.22|
|tsbs-scale100-double-groupby-all@35rps     |      1146.751|        1123.903|                   2.03|
|tsbs-scale100-groupby-orderby-limit@28rps  |        11.760|          11.529|                   2.00|
|tsbs-scale100-high-cpu-1@1816rps           |        45.227|          43.179|                   4.74|
|tsbs-scale100-high-cpu-all@18rps           |      2969.983|        2943.615|                   0.90|
|tsbs-scale100-lastpoint@488rps             |         7.887|           7.742|                   1.87|
|tsbs-scale100-single-groupby-1-1-12@2320rps|         1.891|           1.903|                  -0.63|
|tsbs-scale100-single-groupby-1-1-1@2320rps |         0.836|           0.834|                   0.24|
|tsbs-scale100-single-groupby-1-8-1@4458rps |         8.332|           8.252|                   0.97|
|tsbs-scale100-single-groupby-5-1-12@648rps |         9.321|           9.188|                   1.45|
|tsbs-scale100-single-groupby-5-1-1@648rps  |         1.375|           1.351|                   1.78|
|tsbs-scale100-single-groupby-5-8-1@1158rps |        13.638|          12.266|                  11.19|


# 15 primaries

## Comparison between master and perf.compressed.iterator for metric: Totals.overallQuantiles.all_queries.q50
|                 Test Case                 |Baseline value|Comparison Value|% change (lower-better)|
|-------------------------------------------|--------------|----------------|----------------------:|
|tsbs-scale100-cpu-max-all-1@4139rps        |        10.641|          10.521|                   1.14|
|tsbs-scale100-cpu-max-all-8@588rps         |        35.365|          32.891|                   7.52|
|tsbs-scale100-double-groupby-1@324rps      |        46.653|          43.963|                   6.12|
|tsbs-scale100-double-groupby-5@70rps       |       200.015|         189.943|                   5.30|
|tsbs-scale100-double-groupby-all@35rps     |       371.423|         356.271|                   4.25|
|tsbs-scale100-groupby-orderby-limit@28rps  |        12.820|          12.583|                   1.88|
|tsbs-scale100-high-cpu-1@1816rps           |        39.387|          39.237|                   0.38|
|tsbs-scale100-high-cpu-all@18rps           |      1830.015|        1814.207|                   0.87|
|tsbs-scale100-lastpoint@488rps             |         8.298|           8.310|                  -0.14|
|tsbs-scale100-single-groupby-1-1-12@2320rps|         4.098|           4.069|                   0.71|
|tsbs-scale100-single-groupby-1-1-1@2320rps |         2.928|           2.942|                  -0.48|
|tsbs-scale100-single-groupby-1-8-1@4458rps |         9.540|           9.481|                   0.62|
|tsbs-scale100-single-groupby-5-1-12@648rps |         9.703|           9.674|                   0.30|
|tsbs-scale100-single-groupby-5-1-1@648rps  |         2.226|           2.225|                   0.04|
|tsbs-scale100-single-groupby-5-8-1@1158rps |         9.407|           9.106|                   3.31|


# 30 primaries
## Comparison between master and perf.compressed.iterator for metric: Totals.overallQuantiles.all_queries.q50
|                 Test Case                 |Baseline value|Comparison Value|% change (lower-better)|
|-------------------------------------------|-------------:|---------------:|----------------------:|
|tsbs-scale100-cpu-max-all-1@4139rps        |        15.556|          15.381|                   1.14|
|tsbs-scale100-cpu-max-all-8@588rps         |        33.777|          31.894|                   5.90|
|tsbs-scale100-double-groupby-1@324rps      |        43.913|          41.867|                   4.89|
|tsbs-scale100-double-groupby-5@70rps       |       193.191|         180.871|                   6.81|
|tsbs-scale100-double-groupby-all@35rps     |       353.999|         335.791|                   5.42|
|tsbs-scale100-groupby-orderby-limit@28rps  |        14.348|          14.030|                   2.27|
|tsbs-scale100-high-cpu-1@1816rps           |        44.059|          44.047|                   0.03|
|tsbs-scale100-high-cpu-all@18rps           |      1551.807|        1505.791|                   3.06|
|tsbs-scale100-lastpoint@488rps             |        10.222|          10.240|                  -0.18|
|tsbs-scale100-single-groupby-1-1-12@2320rps|        11.947|          11.750|                   1.68|
|tsbs-scale100-single-groupby-1-1-1@2320rps |         9.762|           9.712|                   0.51|
|tsbs-scale100-single-groupby-1-8-1@4458rps |        14.488|          14.335|                   1.07|
|tsbs-scale100-single-groupby-5-1-12@648rps |        10.885|          10.875|                   0.09|
|tsbs-scale100-single-groupby-5-1-1@648rps  |         4.204|           4.259|                  -1.29|
|tsbs-scale100-single-groupby-5-8-1@1158rps |        13.826|          13.558|                   1.98|